+++ changelog.md
@@ -8,6 +8,7 @@
     app-mode: chrome process outputs to null
     fix: logging server side
     the use of embed window (returning guy class) is now (really) deprecated (wants to simplify)
+    app-mode: disable google translate
 
 0.5.6: (09/04/2020)
     fix: js log was always on. now: it depends if the log is activated or on server side too